update from origin master

+++ flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/ZooKeeperCheckpointIDCounter.java
@@ -23,14 +23,15 @@ import org.apache.curator.framework.recipes.shared.SharedCount;
 import org.apache.curator.framework.recipes.shared.VersionedValue;
 import org.apache.curator.framework.state.ConnectionState;
 import org.apache.curator.framework.state.ConnectionStateListener;
-import org.apache.flink.runtime.jobmanager.RecoveryMode;
+import org.apache.flink.runtime.jobgraph.JobStatus;
+import org.apache.flink.runtime.jobmanager.HighAvailabilityMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
 
 /**
- * {@link CheckpointIDCounter} instances for JobManagers running in {@link RecoveryMode#ZOOKEEPER}.
+ * {@link CheckpointIDCounter} instances for JobManagers running in {@link HighAvailabilityMode#ZOOKEEPER}.
  *
  * <p>Each counter creates a ZNode:
  * <pre>
@@ -70,9 +71,8 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 	 *
 	 * @param client      Curator ZooKeeper client
 	 * @param counterPath ZooKeeper path for the counter. It's sufficient to have a path per-job.
-	 * @throws Exception
 	 */
-	public ZooKeeperCheckpointIDCounter(CuratorFramework client, String counterPath) throws Exception {
+	public ZooKeeperCheckpointIDCounter(CuratorFramework client, String counterPath) {
 		this.client = checkNotNull(client, "Curator client");
 		this.counterPath = checkNotNull(counterPath, "Counter path");
 		this.sharedCount = new SharedCount(client, counterPath, 1);
@@ -91,14 +91,17 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 	}
 
 	@Override
-	public void stop() throws Exception {
+	public void shutdown(JobStatus jobStatus) throws Exception {
 		synchronized (startStopLock) {
 			if (isStarted) {
+				LOG.info("Shutting down.");
 				sharedCount.close();
 				client.getConnectionStateListenable().removeListener(connStateListener);
 
-				LOG.info("Removing {} from ZooKeeper", counterPath);
-				client.delete().deletingChildrenIfNeeded().inBackground().forPath(counterPath);
+				if (jobStatus.isGloballyTerminalState()) {
+					LOG.info("Removing {} from ZooKeeper", counterPath);
+					client.delete().deletingChildrenIfNeeded().inBackground().forPath(counterPath);
+				}
 
 				isStarted = false;
 			}
@@ -115,8 +118,13 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 			}
 
 			VersionedValue<Integer> current = sharedCount.getVersionedValue();
+			int newCount = current.getValue() + 1;
 
-			Integer newCount = current.getValue() + 1;
+			if (newCount < 0) {
+				// overflow and wrap around
+				throw new Exception("Checkpoint counter overflow. ZooKeeper checkpoint counter only supports " +
+						"checkpoints Ids up to " + Integer.MAX_VALUE);
+			}
 
 			if (sharedCount.trySetCount(current, newCount)) {
 				return current.getValue();
@@ -145,7 +153,7 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 	 * Connection state listener. In case of {@link ConnectionState#SUSPENDED} or {@link
 	 * ConnectionState#LOST} we are not guaranteed to read a current count from ZooKeeper.
 	 */
-	private class SharedCountConnectionStateListener implements ConnectionStateListener {
+	private static class SharedCountConnectionStateListener implements ConnectionStateListener {
 
 		private volatile ConnectionState lastState;
